requirements.txt
psycopg[binary]==3.2.3
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
pyarrow==18.1.0
